working version of the core
