+++ user-service/src/main/java/com/ementor/core/core/config/JwtAuthenticationFilter.java
@@ -1,0 +1,61 @@
+/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
+package com.ementor.core.core.config;
+
+import com.ementor.core.core.redis.repo.StoredRedisTokenRepo;
+import com.ementor.core.core.service.JwtService;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+@Component
+@RequiredArgsConstructor
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+	private final JwtService jwtService;
+	private final UserDetailsService userDetailsService;
+	private final StoredRedisTokenRepo storedRedisTokenRepo;
+
+	@Override
+	protected void doFilterInternal(@NonNull HttpServletRequest request,
+			@NonNull HttpServletResponse response,
+			@NonNull FilterChain filterChain) throws ServletException,
+			IOException {
+		if (request.getServletPath()
+			.contains("/api/v1/auth")) {
+			filterChain.doFilter(request, response);
+			return;
+		}
+		final String authHeader = request.getHeader("Authorization");
+		final String jwt;
+		final String userEmail;
+		if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+			filterChain.doFilter(request, response);
+			return;
+		}
+		jwt = authHeader.substring(7);
+		userEmail = jwtService.extractUsername(jwt);
+		if (userEmail != null && SecurityContextHolder.getContext()
+			.getAuthentication() == null) {
+			UserDetails userDetails = this.userDetailsService.loadUserByUsername(userEmail);
+			if (jwtService.isTokenValid(jwt, userDetails)) {
+				UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(userDetails, null,
+						userDetails.getAuthorities());
+				authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+				SecurityContextHolder.getContext()
+					.setAuthentication(authToken);
+			}
+		}
+		filterChain.doFilter(request, response);
+	}
+}
