+++ user-service/src/main/java/com/ementor/core/enums/RoleEnum.java
@@ -1,0 +1,6 @@
+/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
+package com.ementor.core.enums;
+
+public enum RoleEnum {
+	ADMIN, PROFESSOR, STUDENT
+}
