+++ user-service/src/main/java/com/ementor/core/utils/ConstantUtils.java
@@ -1,0 +1,16 @@
+/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
+package com.ementor.core.utils;
+
+//SET HERE WHEN MODIFIED
+
+public class ConstantUtils {
+	public static final String RedisHashName = "StoredRedisToken";
+
+	public static final long port = 49199;
+	public static final String localEnvUrl = "http://localhost:" + port;
+
+	public static final String productionHost = "http://api.ementor.dinca.one";
+
+	public static final String productionService = "service1";
+	public static final String productionEnvUrl = productionHost + "/" + productionService;
+}
