--- user-service/src/main/java/com/ementor/userservice/dto/AuthenticationRequestDTO.java
@@ -1,17 +1,0 @@
-/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
-package com.ementor.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequestDTO {
-
-	private String email;
-	String password;
-}
