--- user-service/src/main/java/com/ementor/userservice/dto/AuthenticationResponseDTO.java
@@ -1,19 +1,0 @@
-/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
-package com.ementor.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponseDTO {
-
-	private String accessToken;
-	private String refreshToken;
-
-	private UserGetDTO userData;
-}
