--- user-service/src/main/java/com/ementor/userservice/dto/RegisterRequestDTO.java
@@ -1,20 +1,0 @@
-/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
-package com.ementor.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequestDTO {
-
-	private String firstName;
-	private String lastName;
-	private String email;
-	private String password;
-	private String phone;
-}
