--- user-service/src/main/java/com/ementor/userservice/enums/RoleEnum.java
@@ -1,6 +1,0 @@
-/* Copyright (C) 2022-2023 Ementor Romania - All Rights Reserved */
-package com.ementor.userservice.enums;
-
-public enum RoleEnum {
-	ADMIN, PROFESSOR, STUDENT
-}
