+++ user-service/src/main/resources/application.properties
@@ -1,10 +1,10 @@
 #server config
-server.port=49200
+server.port=49199
 #ALSO UPDATE IN CONSTANT UTILS TO WORK
 
 
 #flyway and postgresql config
-ementor.datasource.core.url=jdbc:postgresql://127.0.0.1:5432/user_service
+ementor.datasource.core.url=jdbc:postgresql://127.0.0.1:5432/core_service
 ementor.datasource.core.driver-class-name=org.postgresql.Driver
 ementor.datasource.core.username=postgres
 ementor.datasource.core.password=postgres
